Simplifying build process by using a single build script
Previously we had several sets of config files and build scripts for different environment targets. This allowed us to build for CJS, UMD and other different formats.
This didn't make much sense because environments that don't support ESModules probably also do not support WebAssembly. I may change this later because I want to support the most evironments possible but it was hindering development, so it will stay this way for now.